requirements.txt
pymupdf
tqdm
python-dotenv
pillow
langchain
langgraph
langchain-groq  # TODO: make installation with pyproject.toml with intallation options
langchain-openai  # TODO: make installation with pyproject.toml with intallation options

# just for convenience
iprogress
